Moved abstraction and implementation to corresponding folders
