+++ InvoicingApplication/DAL/Abstraction/IRepository.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq.Expressions;
 
-namespace DAL
+namespace DAL.Abstraction
 {
     public interface IRepository<TEntity> where TEntity : class
     {
