+++ InvoicingApplication/DAL/Implementation/Repository.cs
@@ -1,10 +1,11 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using DAL.Abstraction;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 
-namespace DAL
+namespace DAL.Implementation
 {
     public class Repository<TEntity> : IRepository<TEntity> where TEntity : class
     {
