+++ InvoicingApplication/WpfApplication/bin/Debug/netcoreapp3.0/WpfApplication.runtimeconfig.dev.json
@@ -3,8 +3,7 @@
     "additionalProbingPaths": [
       "C:\\Users\\lesze\\.dotnet\\store\\|arch|\\|tfm|",
       "C:\\Users\\lesze\\.nuget\\packages",
-      "C:\\Microsoft\\Xamarin\\NuGet",
-      "C:\\Program Files\\dotnet\\sdk\\NuGetFallbackFolder"
+      "C:\\Microsoft\\Xamarin\\NuGet"
     ]
   }
 }
